benchmarks.Rmd: First report on matrix results
Add an Rmarkdown document for a simple report on the results of the
benchmark tests. This report focuses on showing the distribution of
results for each configuration of each benchmark. Further reports can
be added to focus on other aspects such as the relative performance of
different Snabb versions or hardware platforms.

The idea is that the reports can all be generated by the nix
expression that runs a test campaign. I have not tied this in yet and
am hoping for help from @domenkozar on that :-). For now I created
this report by downloading the CSV file containing the results (of
Hydra evaluation 1797) and writing the document in Rstudio.

Creating this report highlighted a couple of issues with the CSV file
in that version:

The Snabb software version column is N/A. (I would like to make Snabb
version the color for data points but that produced an error with this
CSV.)

The score for all l2fwd results is N/A. Could be a problem with the
nix expression for producing a CSV row?