Added report-full-matrix.Rmd
This draft report compares Snabb versions for performance across a
large set of tests. The overall test results are summarized and
specific test scenarios are also shown individually.

This is not tied in to the build yet but it is intended for use with
the full test matrix and ideally with multple Snabb versions being
compared.

Example rendering here:
http://rpubs.com/lukego/193284

This was developed using the CSV data from Hydra Evaluation 2100:
https://hydra.snabb.co/build/78984