report-by-snabb.Rmd: Tighten filter on CSV rows
Exclude special configurations of iperf/l2fwd. This cuts the 'murren
large' dataset down to the same rows as the 'lugano small' dataset so
that the reports can be compared directly.

Previously we had too many rows getting mixed up in the 'murren large'
dataset that made it look chaotic.